Pi actors

+++ src/main/scala/pi/PiWorker.scala
@@ -11,8 +11,8 @@ object PiWorker {
 
 class PiWorker extends Actor with ActorLogging{
 
-  def calculatePiFor(start: Int, nrOfElements: Int): Double = {
-    var acc = 0.0
+  def calculatePiFor(start: Int, nrOfElements: Int): BigDecimal = {
+    var acc = BigDecimal(0.0)
     for (i ← start until (start + nrOfElements))
       acc += 4.0 * (1 - (i % 2) * 2) / (2 * i + 1)
     acc
